Make Flex use C++ (#45)
* Create FlexLexer.h

* Convert Flex scanner from C to C++. Add scanner unit tests.

* Replace tabs with spaces

+++ src/ast/int.cpp
@@ -6,7 +6,7 @@
 using namespace AST;
 using namespace std;
 
-Int::Int(const yy::location &loc, int value)
+Int::Int(const yy::location &loc, int64_t value)
     : Expression(loc)
     , value{value} {}
